requirements.txt
pyyaml
pyside6
opencv-python
minimalmodbus